app service

+++ src/main/scala/search/solr/client/entity/searchinterface/NiNi.java
@@ -4,32 +4,32 @@ package search.solr.client.entity.searchinterface;
  * Created by soledede on 2016/4/18.
  */
 public class NiNi {
-    private Integer costMillisecond;
-    private Integer costSecond;
+    private Float costMillisecond;
+    private Float costSecond;
     private Object data;
 
     public NiNi() {
     }
 
-    public NiNi(Integer costMillisecond, Integer costSecond, Object data) {
+    public NiNi(Float costMillisecond, Float costSecond, Object data) {
         this.costMillisecond = costMillisecond;
         this.costSecond = costSecond;
         this.data = data;
     }
 
-    public Integer getCostMillisecond() {
+    public Float getCostMillisecond() {
         return costMillisecond;
     }
 
-    public void setCostMillisecond(Integer costMillisecond) {
+    public void setCostMillisecond(Float costMillisecond) {
         this.costMillisecond = costMillisecond;
     }
 
-    public Integer getCostSecond() {
+    public Float getCostSecond() {
         return costSecond;
     }
 
-    public void setCostSecond(Integer costSecond) {
+    public void setCostSecond(Float costSecond) {
         this.costSecond = costSecond;
     }
